delivery/requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.1.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
